feat(logger)
add test for export logs
